Add feature to mark attendance, add FOSS, and improve UI
Now invigilator can mark attendance for the advance test on the given day.
After attendance is submitted, attendees can appear for the test.
And invigilator can close the entry for the test.

Administrator can Add/Edit a FOSS for advance test.

Added tabs for the UI of tests with different status